refactor: remove DTOColumns.php trait
